Move helper function out of migrate into triggers
